requirements.txt
django == 5.0.1
psycopg[binary] == 3.1.17
django-import-export == 3.3.6
